引入RxJava2处理耗时操作，提高代码可读性和方便业务逻辑处理

+++ app/src/main/java/com/konkawise/dtv/utils/TimeUtils.java
@@ -225,7 +225,7 @@ public class TimeUtils {
      *
      * @return 00:00:00
      */
-    public static String getDecimalFormatTime(int seconds) {
+    public static String getDecimalFormatTime(long seconds) {
         if (seconds <= 0) {
             return "00:00:00";
         }
